fix: refactor validation of repo and backup target paths to resolve issue with remote repos
Previously, the repo and backup_target fields of BorgRepo had a union type of DirectoryPath
NewPath. This breaks when trying to validate a linux path on Mac and vice versa. I moved the path validation into a new model_validator that only runs if the hostname for the repo matches the current hostname.

BREAKING CHANGE: Updated type of both path and backup_target fields in BorgRepo from Path to str